test(data-view): split the util suites out and cover the gaps
`packLanes` and `orderByX` were tested from inside timeline.test.tsx,
which had grown to 2,060 lines covering three unrelated subjects. Moves
them to pack-lanes.test.ts and order-by-x.test.ts, with the shared LCG,
FNV-1a digest and item generator in helpers.ts so the recorded goldens
keep being built from byte-identical input.

Pure relocation for the existing 18 — the six inline snapshots are
transcribed verbatim and still match. New coverage, aimed at the paths
nothing reached:

- lanes above 1024, where the free-lane bitmap's summary is two-level and
  every prior test topped out at 400 lanes, leaving the multi-block walk
  in `takeSmallestFree` unexercised. Widths shrink as x grows so high
  lanes free first, forcing the search past an all-zero block-0 summary.
- both sides of each implementation threshold (63 vs 64 items), since
  both utils pick between two algorithms by item count and nothing pinned
  that they agree at the boundary — a mismatch means one extra card
  silently repacks every lane.
- non-finite x and width, which is what surfaced the bug fixed in the
  previous commit.
- gapPx of 0, zero-width items, entirely negative coordinates, and a
  single item.

187 tests pass, up from 176.
