+++ packages/raystack/components/data-view/__tests__/timeline.test.tsx
@@ -20,8 +20,6 @@ import type {
   TimelineActions
 } from '../data-view.types';
 import { useDataView } from '../hooks/useDataView';
-import { orderByX } from '../utils/order-by-x';
-import { packLanes } from '../utils/pack-lanes';
 import { buildAxis, createTimeScale, toTimestamp } from '../utils/time-scale';
 
 beforeAll(() => {
@@ -40,277 +38,6 @@ beforeAll(() => {
 
 afterEach(() => {
   vi.restoreAllMocks();
-});
-
-/* ─────────────────── characterisation: lane assignment ───────────────────
-   Goldens pinning `packLanes` output at sizes too large to hand-write. They
-   were recorded from the implementation before the O(n) rewrite, so a digest
-   mismatch means lane assignment moved — which is a visible change: a card's
-   lane is its vertical position on the canvas, and `context.laneIndex` is
-   public API through `renderCard`.
-
-   Digests rather than 500-element literals: the arrays are only ever compared,
-   never read, and an unreadable wall of numbers hides what the test is for. */
-
-/** Seeded LCG — a failing case has to be reproducible. */
-function seededRandom(seed: number) {
-  let state = seed;
-  return () => {
-    state = (state * 1664525 + 1013904223) >>> 0;
-    return state / 0x100000000;
-  };
-}
-
-/** FNV-1a over the decimal text, so [1, 23] and [12, 3] can't collide. */
-function digest(values: readonly number[]): string {
-  let hash = 0x811c9dc5;
-  for (const value of values) {
-    const text = `${value},`;
-    for (let i = 0; i < text.length; i++) {
-      hash ^= text.charCodeAt(i);
-      hash = Math.imul(hash, 0x01000193) >>> 0;
-    }
-  }
-  return hash.toString(16).padStart(8, '0');
-}
-
-const randomItems = (seed: number, count: number) => {
-  const random = seededRandom(seed);
-  return Array.from({ length: count }, () => ({
-    x: Math.round(random() * 10000),
-    width: Math.round(random() * 200)
-  }));
-};
-
-describe('packLanes (characterisation)', () => {
-  it('pins lane assignment for randomly spread items', () => {
-    const { lanes, laneCount } = packLanes(randomItems(42, 500));
-    expect({ laneCount, lanes: digest(lanes) }).toMatchInlineSnapshot(`
-      {
-        "laneCount": 13,
-        "lanes": "908c8fec",
-      }
-    `);
-  });
-
-  it('pins lane assignment for a second, denser spread', () => {
-    const { lanes, laneCount } = packLanes(randomItems(7, 500), 4);
-    expect({ laneCount, lanes: digest(lanes) }).toMatchInlineSnapshot(`
-      {
-        "laneCount": 12,
-        "lanes": "e1515f6a",
-      }
-    `);
-  });
-
-  it('pins lane assignment when every item overlaps every other', () => {
-    // No lane is ever reusable, so lane count tracks item count — the shape
-    // that makes the lane scan quadratic.
-    const items = Array.from({ length: 400 }, (_, i) => ({
-      x: i,
-      width: 10000
-    }));
-    const { lanes, laneCount } = packLanes(items);
-    expect({ laneCount, lanes: digest(lanes) }).toMatchInlineSnapshot(`
-      {
-        "laneCount": 400,
-        "lanes": "732748cb",
-      }
-    `);
-  });
-
-  it('pins lane assignment when items cluster on a single x', () => {
-    const items = Array.from({ length: 200 }, (_, i) => ({
-      x: 500,
-      width: i % 3
-    }));
-    const { lanes, laneCount } = packLanes(items);
-    expect({ laneCount, lanes: digest(lanes) }).toMatchInlineSnapshot(`
-      {
-        "laneCount": 200,
-        "lanes": "23ec0f1f",
-      }
-    `);
-  });
-
-  it('pins lane assignment across the gap boundary', () => {
-    // Alternates releases landing exactly on the gap boundary (reusable) with
-    // ones a pixel short (not) — the comparison most at risk from a rewrite.
-    const items = Array.from({ length: 300 }, (_, i) => ({
-      x: i * 108 + (i % 2),
-      width: 100
-    }));
-    const { lanes, laneCount } = packLanes(items);
-    expect({ laneCount, lanes: digest(lanes) }).toMatchInlineSnapshot(`
-      {
-        "laneCount": 2,
-        "lanes": "05a0dca4",
-      }
-    `);
-  });
-
-  it('pins lane assignment for items sharing exact edges', () => {
-    // Ties on x, where assignment depends on the visit order the sort gives.
-    const items = Array.from({ length: 120 }, (_, i) => ({
-      x: (i % 4) * 250,
-      width: 100 + (i % 7) * 10
-    }));
-    const { lanes, laneCount } = packLanes(items);
-    expect({ laneCount, lanes: digest(lanes) }).toMatchInlineSnapshot(`
-      {
-        "laneCount": 30,
-        "lanes": "34e5e40d",
-      }
-    `);
-  });
-});
-
-/* ─────────────────────────── utils: packLanes ────────────────────────── */
-
-describe('packLanes', () => {
-  it('returns no lanes for empty input', () => {
-    expect(packLanes([])).toEqual({ lanes: [], laneCount: 0 });
-  });
-
-  it('packs non-overlapping items into the same lane', () => {
-    const { lanes, laneCount } = packLanes([
-      { x: 0, width: 100 },
-      { x: 120, width: 50 }
-    ]);
-    expect(lanes).toEqual([0, 0]);
-    expect(laneCount).toBe(1);
-  });
-
-  it('opens a new lane for overlapping items', () => {
-    const { lanes, laneCount } = packLanes([
-      { x: 0, width: 100 },
-      { x: 50, width: 100 }
-    ]);
-    expect(lanes).toEqual([0, 1]);
-    expect(laneCount).toBe(2);
-  });
-
-  it('respects the gap: items closer than gapPx do not share a lane', () => {
-    // First ends at 100; second starts at 104 < 100 + 8 → new lane.
-    const tight = packLanes([
-      { x: 0, width: 100 },
-      { x: 104, width: 50 }
-    ]);
-    expect(tight.lanes).toEqual([0, 1]);
-    // Exactly at the gap boundary → same lane.
-    const exact = packLanes([
-      { x: 0, width: 100 },
-      { x: 108, width: 50 }
-    ]);
-    expect(exact.lanes).toEqual([0, 0]);
-  });
-
-  it('assigns lanes by ascending x regardless of input order', () => {
-    const { lanes, laneCount } = packLanes([
-      { x: 220, width: 60 }, // fits after the first item
-      { x: 0, width: 100 },
-      { x: 50, width: 100 } // overlaps the first → lane 1
-    ]);
-    expect(lanes).toEqual([0, 0, 1]);
-    expect(laneCount).toBe(2);
-  });
-
-  it('matches a lane scan across many random inputs', () => {
-    // The goldens above pin six fixed shapes; this sweeps far more input than
-    // snapshots can carry, and reports the mismatched lane directly rather
-    // than as a changed hash. The oracle is the pre-rewrite implementation,
-    // transcribed — every input here is past the size where packLanes switches
-    // from that scan to the sweep.
-    function packByScanReference(
-      items: { x: number; width: number }[],
-      gapPx = 8
-    ) {
-      const order = items
-        .map((_, i) => i)
-        .sort((a, b) => items[a].x - items[b].x || a - b);
-      const laneEnds: number[] = [];
-      const lanes = new Array<number>(items.length).fill(0);
-      for (const index of order) {
-        const item = items[index];
-        let lane = laneEnds.findIndex(end => end + gapPx <= item.x);
-        if (lane === -1) {
-          lane = laneEnds.length;
-          laneEnds.push(0);
-        }
-        laneEnds[lane] = item.x + item.width;
-        lanes[index] = lane;
-      }
-      return { lanes, laneCount: laneEnds.length };
-    }
-
-    for (let seed = 1; seed <= 25; seed++) {
-      const items = randomItems(seed, 300);
-      expect(packLanes(items)).toEqual(packByScanReference(items));
-    }
-  });
-});
-
-/* ─────────────────────────── utils: orderByX ─────────────────────────── */
-
-describe('orderByX', () => {
-  const sortedXs = (items: { x: number }[]) =>
-    Array.from(orderByX(items), index => items[index].x);
-
-  it('returns an empty order for empty input', () => {
-    expect(Array.from(orderByX([]))).toEqual([]);
-  });
-
-  it('orders ascending by x, breaking ties by input order', () => {
-    const items = [{ x: 30 }, { x: 10 }, { x: 30 }, { x: -5 }];
-    expect(Array.from(orderByX(items))).toEqual([3, 1, 0, 2]);
-  });
-
-  /* Past 64 items orderByX swaps its comparison sort for a counting sort, so
-     the rest of these run above that threshold. */
-
-  it('matches a comparison sort on randomly spread items', () => {
-    const random = seededRandom(7);
-    for (let round = 0; round < 20; round++) {
-      const items = Array.from({ length: 500 }, () => ({
-        x: Math.round((random() - 0.5) * 20000)
-      }));
-      const expected = items
-        .map((_, i) => i)
-        .sort((a, b) => items[a].x - items[b].x || a - b);
-      expect(Array.from(orderByX(items))).toEqual(expected);
-    }
-  });
-
-  it('keeps input order when every item shares one x', () => {
-    // Zero extent — nothing to bucket by, and the tie-break is input order.
-    const items = Array.from({ length: 100 }, () => ({ x: 42 }));
-    expect(Array.from(orderByX(items))).toEqual(
-      Array.from({ length: 100 }, (_, i) => i)
-    );
-  });
-
-  it('sorts a bucket deeper than the insertion-sort cutoff', () => {
-    // 80 items on one x land in a single bucket, past the depth where that
-    // bucket hands off to a comparison sort.
-    const items = [
-      ...Array.from({ length: 80 }, () => ({ x: 100 })),
-      ...Array.from({ length: 40 }, (_, i) => ({ x: 900 - i }))
-    ];
-    expect(sortedXs(items)).toEqual(
-      items.map(item => item.x).sort((a, b) => a - b)
-    );
-  });
-
-  it('orders items clustered at both ends of the extent', () => {
-    // A near-empty middle makes the uniform bucket split maximally uneven.
-    const items = [
-      ...Array.from({ length: 60 }, (_, i) => ({ x: i / 1000 })),
-      ...Array.from({ length: 60 }, (_, i) => ({ x: 10000 + i / 1000 }))
-    ];
-    expect(sortedXs(items)).toEqual(
-      items.map(item => item.x).sort((a, b) => a - b)
-    );
-  });
 });
 
 /* ─────────────────────────── utils: time scale ───────────────────────── */
